requirement.txt
basicsr==1.4.2
datasets==2.9.0
einops==0.6.0
facexlib==0.2.5
fake-factory==9999.9.9
gdown==4.6.4
gpytorch==1.9.0
grpcio==1.51.3
huggingface-hub==0.11.1
hydra-core==1.2.0
ipywidgets==8.0.3
lpips==0.1.4
natsort==8.2.0
opencv-python==4.5.5.64
pandas==1.4.2
Pillow==9.0.1
timm==0.6.12
tokenizers==0.13.2
torch==1.13.0
torchaudio==0.10.2
torchmetrics==0.11.1
torchvision==0.11.3
tqdm==4.64.0
transformers==4.26.1
